feat: 에디터 방향키 이벤트, 이벤트 함수화
